Move flag definitions into extra files
